requirements.txt
# Core Dependencies for Protein Structure Prediction App
# =====================================================

# Streamlit and UI
streamlit>=1.28.0
plotly>=5.15.0

# Machine Learning and Transformers
torch>=2.0.0
transformers>=4.35.0

# Scientific Computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Bioinformatics
biopython>=1.81

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
networkx>=3.1

# 3D Structure Visualization (Optional but Recommended)
py3dmol>=2.0.0

# Additional utilities
warnings

# Optional: For GPU support
# Install with: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118